stacks: refactor shared functionality in prep for removed blocks
