+++ internal/stacks/stackruntime/internal/stackeval/provider_config.go
@@ -82,7 +82,7 @@ func (p *ProviderConfig) ProviderArgs(ctx context.Context, phase EvalPhase) cty.
 	return v
 }
 
-func CheckProviderInLockfile(locks depsfile.Locks, providerType *ProviderType, declRange tfdiags.SourceRange) (diags tfdiags.Diagnostics) {
+func CheckProviderInLockfile(locks depsfile.Locks, providerType *ProviderType, declRange *hcl.Range) (diags tfdiags.Diagnostics) {
 	if !depsfile.ProviderIsLockable(providerType.Addr()) {
 		return diags
 	}
@@ -95,7 +95,7 @@ func CheckProviderInLockfile(locks depsfile.Locks, providerType *ProviderType, d
 				"Provider %q is not in the lockfile. This provider must be in the lockfile to be used in the configuration. Please run `tfstacks providers lock` to update the lockfile and run this operation again with an updated configuration.",
 				providerType.Addr(),
 			),
-			Subject: declRange.ToHCL().Ptr(),
+			Subject: declRange,
 		})
 	}
 	return diags
@@ -114,7 +114,7 @@ func (p *ProviderConfig) CheckProviderArgs(ctx context.Context, phase EvalPhase)
 			if depLocks != nil {
 				// Check if the provider is in the lockfile,
 				// if it is not we can not read the provider schema
-				lockfileDiags := CheckProviderInLockfile(*depLocks, providerType, decl.DeclRange)
+				lockfileDiags := CheckProviderInLockfile(*depLocks, providerType, decl.DeclRange.ToHCL().Ptr())
 				if lockfileDiags.HasErrors() {
 					return cty.DynamicVal, lockfileDiags
 				}
